a lot of work on abinit input reading
also changed float32 to float64 default
added alot of qe stuff too

+++ test/file_processing_tests.jl
@@ -41,7 +41,7 @@ wan_test = read_wannier_input(joinpath(@__DIR__,"../assets/inputs/wannier/wan.wi
 @test length(get_data(wan_test,:kpoints)) == prod(get_flag(wan_test,:mp_grid))
 @test length(get_data(wan_test,:atoms_frac)) == 2
 @test get_flag(wan_test,:write_rmn)
-@test get_data(wan_test,:atoms_frac)[:Te] == [Point3D{Float32}(0.52325284,0.52325284,0.52325284)]
+@test get_data(wan_test,:atoms_frac)[:Te] == [Point3D{Float64}(0.52325284,0.52325284,0.52325284)]
 
 test_filename = joinpath(@__DIR__,"../assets/inputs/wannier/wan_test.win")
 write_input(wan_test,test_filename)
